Refactor and add User-Download functionality
User-Download can be used when a mod cannot be directly downloaded due
to licensing issues or some other reason. The modlist can contain a URL
for the user to visit in order to download the mod and it will tell the
user where to save it in order for it to be moved to the cache.

Another advantage of using a cache here is that the user only needs to
do this manual download once.